ml-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
pinecone-client==3.0.0
openai==1.3.0
textblob==0.17.1
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
